Updated Vaadin version to 7.0.3. Fixed bug in examples which caused the examples to fail with Vaadin 7.0.3. Moved root pom to root folder. Added root pom as parent to child poms and put shared things to the parent pom. Updated checkstyle checks.xml. Reformatted code.Updated maven plugin versions. Switched to java 7.

+++ vaadin-lazyquerycontainer-jpa-example/src/main/java/org/vaadin/addons/lazyquerycontainer/example/jpa/VaadinApplication.java
@@ -155,7 +155,7 @@ public class VaadinApplication extends UI implements ClickListener {
         entityContainer.addContainerProperty("delta", String.class, "", false, true);
         entityContainer.addContainerProperty(LazyQueryView.DEBUG_PROPERTY_ID_QUERY_INDEX, Integer.class, 0, true, false);
         entityContainer.addContainerProperty(LazyQueryView.DEBUG_PROPERTY_ID_BATCH_INDEX, Integer.class, 0, true, false);
-        entityContainer.addContainerProperty(LazyQueryView.DEBUG_PROPERTY_ID_BATCH_QUERY_TIME, Integer.class, 0, true,
+        entityContainer.addContainerProperty(LazyQueryView.DEBUG_PROPERTY_ID_BATCH_QUERY_TIME, Long.class, 0, true,
                 false);
 
         Task entity = null;
